requirements.txt
pytesseract
Pillow
googletrans==4.0.0rc1
pystray
keyboard
pywin32
